Create led patterns.c
This C code is intended to be flashed into a STM32 Nucleo board. IT allows LEDs connected to GPIOs of the board to blink with patterns. The code allows for 6 different modes. The modes can be switched using 3 slider buttons making possible 8 combinations. The remaining two combinations allow for a combination of two modes.

The project link on Wokwi is:

https://wokwi.com/projects/392038492285546497